Navbar Collapse on Resize Functionality #6
When the window is resized to something smaller, the navbar collapses to display a hamburger icon with links to pages. This is helpful for better mobile compatibility. Some code cleanup was done, for example, there was a strange space between the navbar and the html body for the homepage, which was fixed.